Authentication and file handling

+++ db/db_user.py
@@ -2,6 +2,8 @@ from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from db.models import DbUser
 from db.hash import Hash
+from fastapi import HTTPException, status
+
 def create_user(db:Session,request:UserBase):
     new_user = DbUser(
         username = request.username,
@@ -17,10 +19,15 @@ def get_all_users(db:Session):
     return db.query(DbUser).all()
 
 def get_user(db:Session, id:int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
+    return 
 
-def update_user(db:Session,id:int, request:UserBase):
+def update_user(db:Session,id:int, request:UserBase): 
     user=db.query(DbUser).filter(DbUser.id==id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
     user.update({
         DbUser.username:request.username,
         DbUser.email:request.email,
@@ -31,6 +38,9 @@ def update_user(db:Session,id:int, request:UserBase):
 
 def delete_user(db:Session, id:int):
     user=db.query(DbUser).filter(DbUser.id == id).first()
+    if not delete_user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
+
     #Handle any exceptions
     db.delete(user)
     db.commit()
